+++ lib/mongoid/relations/targets/enumerable.rb
@@ -218,7 +218,10 @@ module Mongoid
         #
         # @since 2.1.0
         def first
-          matching_document(:first)
+          _loaded.try(:values).try(:first) ||
+            _added[(ul = _unloaded.try(:first)).try(:id)] ||
+            ul ||
+            _added.values.try(:first)
         end
 
         # Initialize the new enumerable either with a criteria or an array.
@@ -299,7 +302,10 @@ module Mongoid
         #
         # @since 2.1.0
         def last
-          matching_document(:last)
+          _added.values.try(:last) ||
+            _loaded.try(:values).try(:last) ||
+            _added[(ul = _unloaded.try(:last)).try(:id)] ||
+            ul
         end
 
         # Loads all the documents in the enumerable from the database.
@@ -456,13 +462,6 @@ module Mongoid
 
         def method_missing(name, *args, &block)
           entries.send(name, *args, &block)
-        end
-
-        def matching_document(location)
-          _loaded.try(:values).try(location) ||
-            _added[(ul = _unloaded.try(location)).try(:id)] ||
-            ul ||
-            _added.values.try(location)
         end
 
         def unloaded_documents